update JS maps

+++ src/Juniper.Root/Processes/CopyCommand.cs
@@ -5,6 +5,7 @@ namespace Juniper.Processes
         private readonly FileInfo from;
         private readonly FileInfo to;
         private readonly bool warnIfNotExists;
+        private readonly (string, string)? replacement;
         private DateTime? lastWriteTime;
 
         public CopyCommand(string name, FileInfo from, FileInfo to, bool warnIfNotExists)
@@ -13,6 +14,12 @@ namespace Juniper.Processes
             this.from = from;
             this.to = to;
             this.warnIfNotExists = warnIfNotExists;
+        }
+
+        public CopyCommand(string name, FileInfo from, FileInfo to, bool warnIfNotExists, (string, string) replacement)
+            : this(name, from, to, warnIfNotExists)
+        {
+            this.replacement = replacement;
         }
 
         public override Task RunAsync()
@@ -40,7 +47,29 @@ namespace Juniper.Processes
             else
             {
                 to.Directory?.Create();
-                File.Copy(from.FullName, to.FullName, true);
+                if (replacement is null)
+                {
+                    File.Copy(from.FullName, to.FullName, true);
+                }
+                else
+                {
+                    var (oldValue, newValue) = replacement.Value;
+                    var lines = File.ReadAllLines(from.FullName);
+                    if (lines.Length > 0)
+                    {
+                        var index = lines.Length - 1;
+                        while (index > 0 && lines[index].Length == 0)
+                        {
+                            --index;
+                        }
+                        if (index > 0)
+                        {
+                            OnInfo($"Replacing {oldValue} with {newValue} in '{lines[index]}'");
+                            lines[index] = lines[index].Replace(oldValue, newValue);
+                        }
+                    }
+                    File.WriteAllLines(to.FullName, lines);
+                }
                 lastWriteTime = from.LastWriteTime;
                 OnInfo($"Copied! {fromRel} -> {toRel}");
             }
